Refactor k-field
