+++ panel/src/components/Forms/Field/CheckboxesField.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-import Field from "../Field.vue";
+import Field from "@/ui/components/Field.vue";
 import Input from "../Input.vue";
 import CheckboxesInput from "../Input/CheckboxesInput.vue";
