+++ panel/src/components/Forms/Field/DateField.vue
@@ -32,7 +32,7 @@
 </template>
 
 <script>
-import Field from "../Field.vue";
+import Field from "@/ui/components/Field.vue";
 import Input from "../Input.vue";
 import DateTimeInput from "../Input/DateTimeInput.vue";
