+++ panel/src/components/Forms/Field/EmailField.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import Field from "../Field.vue";
+import Field from "@/ui/components/Field.vue";
 import Input from "../Input.vue";
 import EmailInput from "../Input/EmailInput.vue";
