+++ panel/src/components/Forms/Field/TextField.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-import Field from "../Field.vue";
+import Field from "@/ui/components/Field.vue";
 import Input from "../Input.vue";
 import TextInput from "../Input/TextInput.vue";
