+++ panel/src/components/Forms/Field/ToggleField.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import Field from "../Field.vue";
+import Field from "@/ui/components/Field.vue";
 import Input from "../Input.vue";
 import ToggleInput from "../Input/ToggleInput.vue";
