+++ panel/src/components/Forms/Field/UrlField.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import Field from "../Field.vue";
+import Field from "@/ui/components/Field.vue";
 import Input from "../Input.vue";
 import UrlInput from "../Input/UrlInput.vue";
